Move CpuProfile into arch/src/lib.rs
